# openapi.yaml
# Copyright 2024 Function Stream Org.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

openapi: 3.0.0
info:
  title: Function Stream API
  version: 0.1.0
servers:
  - url: http://localhost:7300
    description: Local development server
paths:
  /api/v1/functions:
    get:
      summary: Returns a list of functions
      responses:
        '200':
          description: A JSON array of function names
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                  example: "test-func"
  /api/v1/function/{function_name}:
    post:
      summary: Starts a function
      parameters:
        - in: path
          name: function_name
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Function'
      responses:
        '200':
          description: Function started successfully
        '400':
          description: Bad request, failed to parse function definition or body is empty
        '409':
          description: Conflict, function already exists
    delete:
      summary: Deletes a function
      parameters:
        - in: path
          name: function_name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Function deleted successfully
        '404':
          description: Function not found
  /api/v1/produce/{queue_name}:
    put:
      summary: Produces an event to a queue
      parameters:
        - in: path
          name: queue_name
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: string
      responses:
        '200':
          description: Event produced successfully
        '400':
          description: Bad request, failed to parse event or body is empty
  /api/v1/consume/{queue_name}:
    get:
      summary: Consumes an event from a queue
      parameters:
        - in: path
          name: queue_name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Event consumed successfully
          content:
            application/json:
              schema:
                type: string
components:
  schemas:
    Function:
      type: object
      properties:
        name:
          type: string
        archive:
          type: string
        inputs:
          type: array
          items:
            type: string
        output:
          type: string
        replicas:
          type: integer
        config:
          type: object
          additionalProperties:
            type: string
      required:
        - archive
        - inputs
        - output
      example:
        name: "test-func"
        archive: "/path/to/func.wasm"
        inputs: [ "input1", "input2" ]
        output: "output"
        config: { "key": "value" }